requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
beautifulsoup4
arxiv
pymupdf
wikipedia
bs4
langchain-text-splitters
langchain-openai
chromadb
SENTENCE-TRANSFORMERS
langchain_huggingface
faiss-cpu
chromadb
langchain_chroma
streamlit
#langchain_chroma-0.2.4